#8562 PlanConfig fixed, messages for DB2TableChooser moved from ext.db2 to ext.db2.ui
