+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources.properties
@@ -90,3 +90,9 @@ edit_db2_index_manager_dialog_title=Create index
 editors_db2_application_editor_title_force_application=Force Application
 editors_db2_application_editor_action_force=Force
 editors_db2_application_editor_confirm_action=Are you sure you want to "force" application "{1}"?
+
+dialog_explain_choose_tablespace=Choose a Tablespace
+dialog_explain_choose_tablespace_tablespace=Tablespace
+dialog_explain_ask_to_create=Tables for EXPLAIN not found within current authorization ID (%s) nor in SYSTOOLS.\n\nDo you want DBeaver to create new EXPLAIN tables?
+dialog_explain_no_tables=Explain tables missing
+dialog_explain_no_tablespace_found_title=No Tablespace
