+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_de.properties
@@ -69,3 +69,10 @@ edit_db2_index_manager_dialog_title = Index erstellen
 editors_db2_application_editor_action_force = Erzwingen
 editors_db2_application_editor_confirm_action = Sind Sie sicher, dass Sie die Anwendung "{1}" "erzwingen" wollen?
 editors_db2_application_editor_title_force_application = Erzwingen der Anwendung
+
+dialog_explain_ask_to_create = Tabellen f\u00FCr EXPLAIN, die weder in der aktuellen Berechtigungs-ID (%s) noch in SYSTOOLS enthalten sind.\r\n\r\nM\u00F6chten Sie, dass DBeaver neue EXPLAIN-Tabellen erstellt?
+dialog_explain_choose_tablespace = Einen Tablespace ausw\u00E4hlen
+dialog_explain_choose_tablespace_tablespace = Tablespace
+dialog_explain_no_tables = Fehlende Tabellen erkl\u00E4ren
+dialog_explain_no_tablespace_found = Es wurde kein verwendbarer Tablespace gefunden, um Explain-Tabellen zu erstellen.
+dialog_explain_no_tablespace_found_title = Kein Tablespace
