+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_ru.properties
@@ -83,3 +83,11 @@ edit_db2_index_manager_dialog_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \
 editors_db2_application_editor_title_force_application=\u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E
 editors_db2_application_editor_action_force=\u041F\u0440\u0435\u0440\u0432\u0430\u0442\u044C
 editors_db2_application_editor_confirm_action=\u0412\u044B \u0443\u0432\u0435\u0440\u0435\u043D\u044B, \u0447\u0442\u043E \u0445\u043E\u0442\u0438\u0442\u0435 \u043F\u0440\u0435\u0440\u0432\u0430\u0442\u044C \u0441\u0435\u0441\u0441\u0438\u044E "{1}"?
+
+dialog_explain_choose_tablespace=\u0412\u044B\u0431\u0435\u0440\u0438\u0442\u0435 \u0442\u0430\u0431\u043B\u0438\u0447\u043D\u043E\u0435 \u043F\u0440\u043E\u0441\u0442\u0440\u0430\u043D\u0441\u0442\u0432\u043E
+dialog_explain_choose_tablespace_tablespace=\u0422\u0430\u0431\u043B\u0438\u0447\u043D\u043E\u0435 \u043F\u0440\u043E\u0441\u0442\u0440\u0430\u043D\u0441\u0442\u0432\u043E
+dialog_explain_no_tables=Explain \u0442\u0430\u0431\u043B\u0438\u0446\u044B \u043E\u0442\u0441\u0443\u0442\u0441\u0442\u0432\u0443\u044E\u0442
+dialog_explain_no_tablespace_found=\u0414\u043B\u044F \u0441\u043E\u0437\u0434\u0430\u043D\u0438\u044F Explain \u0442\u0430\u0431\u043B\u0438\u0446 \u043D\u0435 \u043D\u0430\u0439\u0434\u0435\u043D\u043E \u043D\u0438 \u043E\u0434\u043D\u043E\u0433\u043E \u043F\u0440\u0438\u0433\u043E\u0434\u043D\u043E\u0433\u043E \u0442\u0430\u0431\u043B\u0438\u0447\u043D\u043E\u0433\u043E \u043F\u0440\u043E\u0441\u0442\u0440\u0430\u043D\u0441\u0442\u0432\u0430
+dialog_explain_no_tablespace_found_title=No Tablespace
+dialog_explain_no_tables_found_ex=EXPLAIN \u0442\u0430\u0431\u043B\u0438\u0446\u044B \u043D\u0435 \u043D\u0430\u0439\u0434\u0435\u043D\u044B. \u0417\u0430\u043F\u0440\u043E\u0441 \u043D\u0435 \u043C\u043E\u0436\u0435\u0442 \u0431\u044B\u0442\u044C \u0432\u044B\u043F\u043E\u043B\u043D\u0435\u043D.
+dialog_explain_ask_to_create=\u0422\u0430\u0431\u043B\u0438\u0446\u044B \u0434\u043B\u044F EXPLAIN \u043D\u0435 \u043D\u0430\u0439\u0434\u0435\u043D\u044B \u043D\u0438 \u0432 \u0442\u0435\u043A\u0443\u0449\u0435\u043C id \u0430\u0432\u0442\u043E\u0440\u0438\u0437\u0430\u0446\u0438\u0438 (%s), \u043D\u0438 \u0432 SYSTOOLS.\n\n\u0425\u043E\u0442\u0438\u0442\u0435, \u0447\u0442\u043E\u0431\u044B DBeaver \u0441\u043E\u0437\u0434\u0430\u043B \u043D\u043E\u0432\u044B\u0435 \u0442\u0430\u0431\u043B\u0438\u0446\u044B EXPLAIN?
