+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Messages.java
@@ -33,16 +33,8 @@ public class DB2Messages extends NLS {
         // Pure Utility Class
     }
 
-    public static String dialog_explain_choose_tablespace;
-    public static String dialog_explain_choose_tablespace_tablespace;
     public static String dialog_explain_no_tables;
-    public static String dialog_explain_no_tablespace_found_title;
     public static String dialog_explain_no_tables_found_ex;
-    public static String dialog_explain_ask_to_create;
-
-    public static String editors_db2_application_editor_title_force_application;
-    public static String editors_db2_application_editor_action_force;
-    public static String editors_db2_application_editor_confirm_action;
 
     public static String no_ddl_for_system_tables;
     public static String no_ddl_for_nicknames;
