+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Resources.properties
@@ -1,10 +1,6 @@
-dialog_explain_choose_tablespace=Choose a Tablespace
-dialog_explain_choose_tablespace_tablespace=Tablespace
 dialog_explain_no_tables=Explain tables missing
 dialog_explain_no_tablespace_found=No usable Tablespace has been found to create Explain tables
-dialog_explain_no_tablespace_found_title=No Tablespace
 dialog_explain_no_tables_found_ex=EXPLAIN tables not found. Query cannot be explained.
-dialog_explain_ask_to_create=Tables for EXPLAIN not found within current authorization ID (%s) nor in SYSTOOLS.\n\nDo you want DBeaver to create new EXPLAIN tables?
 
 no_ddl_for_system_tables=-- DDL is not available for System Tables
 no_ddl_for_nicknames=-- DDL is not available for Nicknames
