+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Resources_de.properties
@@ -1,11 +1,6 @@
 #Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
-dialog_explain_ask_to_create = Tabellen f\u00FCr EXPLAIN, die weder in der aktuellen Berechtigungs-ID (%s) noch in SYSTOOLS enthalten sind.\r\n\r\nM\u00F6chten Sie, dass DBeaver neue EXPLAIN-Tabellen erstellt?
-dialog_explain_choose_tablespace = Einen Tablespace ausw\u00E4hlen
-dialog_explain_choose_tablespace_tablespace = Tablespace
 dialog_explain_no_tables = Fehlende Tabellen erkl\u00E4ren
 dialog_explain_no_tables_found_ex = EXPLAIN-Tabellen nicht gefunden. Die Abfrage kann nicht beantwortet werden.
-dialog_explain_no_tablespace_found = Es wurde kein verwendbarer Tablespace gefunden, um Explain-Tabellen zu erstellen.
-dialog_explain_no_tablespace_found_title = Kein Tablespace
 
 no_ddl_for_nicknames = -- DDL ist f\u00FCr Nicknames nicht verf\u00FCgbar.
 no_ddl_for_nonsql_routines = -- DDL ist nicht verf\u00FCgbar f\u00FCr nicht-SQL-basierte Routinen.
